pyproject.toml
[project]
name = "ai-school-rag"
version = "0.1.0"
description = "RAG pipeline project"
authors = [
    {name = "Stanislav Zhurich",email = "stanislav.zhurich@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14.1 || >3.14.1,<3.15"
dependencies = [
    "openai (>=2.30.0,<3.0.0)",
    "chromadb (>=1.5.5,<2.0.0)",
    "kagglehub (>=1.0.0,<2.0.0)",
    "pdfplumber (>=0.11.0,<1.0.0)",
    "pypdf (>=5.0.0,<6.0.0)",
    "pytesseract (>=0.3.13,<1.0.0)",
    "pillow (>=11.0.0,<12.0.0)",
    "transformers (>=5.4.0,<6.0.0)",
    "torch (>=2.11.0,<3.0.0)",
    "torchvision (>=0.26.0,<0.27.0)",
    "pymupdf (>=1.27.2.2,<2.0.0.0)",
    "openai (>=2.30.0,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "streamlit (>=1.40.0,<2.0.0)",
    "plotly (>=5.0.0,<6.0.0)",
    "pandas (>=2.0.0,<3.0.0)",
    "ragas (>=0.2.0,<0.3.0)",
    "langchain-openai (>=0.3.0,<0.4.0)",
    "datasets (>=4.8.4,<5.0.0)"
]


[tool.poetry]
packages = [{include = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
